remove some files and add script for db setup
remove some files and add script for db setup
