+++ database-setup/create-db.sql
@@ -1,0 +1,3 @@
+CREATE DATABASE database_name;
+
+GRANT CONNECT ON DATABASE database_name TO user_name;
