+++ database-setup/drop-db.sql
@@ -1,0 +1,1 @@
+DROP DATABASE IF EXISTS database_name;
