+++ database-setup/drop-role.sql
@@ -1,0 +1,1 @@
+DROP ROLE IF EXISTS user_name;
