+++ src/main/java/org/folio/marccat/config/Global.java
@@ -41,8 +41,8 @@ public abstract class Global {
   public static final String TEXT_CODE = "t";
   public static final String VIDEO_RECORDING = "v";
   public static final String UNSPECIFIED = "z";
-  public static final List<String> FIXED_FIELDS = new ArrayList<>(Arrays.asList("000", "001", "005", "006", "007", "008"));
-  public static final List<String> MANDATORY_FIELDS = new ArrayList<>(Arrays.asList("000", "001", "008", "040"));
+  public static final List<String> FIXED_FIELDS = Arrays.asList("000", "001", "005", "006", "007", "008");
+  public static final List<String> MANDATORY_FIELDS = Arrays.asList("000", "001", "008", "040");
   public static final String HEADER_TYPE_LABEL = "HEADER_TYPE";
   public static final String FORM_OF_MATERIAL_LABEL = "FORM_OF_MATERIAL";
   public static final String MATERIAL_TYPE_CODE_LABEL = "MATERIAL_TYPE_CODE";
@@ -457,27 +457,27 @@ public abstract class Global {
 
   public static final char BIBLIOGRAPHIC_INDICATOR_NOT_NUMERIC = 'S';
 
-  public static final List<String> NAMES =  new ArrayList<>(Arrays.asList("100", "110", "111"	));
+  public static final List<String> NAMES = Arrays.asList("100", "110", "111"	);
 
-  public static final List<String> NAMES_D = new ArrayList<>(Arrays.asList("110", "111", "710", "711"));
+  public static final List<String> NAMES_D = Arrays.asList("110", "111", "710", "711");
 
-  public static final List<String> NAMES_E = new ArrayList<>(Arrays.asList("100", "110", "120", "121", "240", "243", "400", "410", "600", "610", "700", "710", "720", "721", "800", "810", "900", "910", "980", "981"));
+  public static final List<String> NAMES_E = Arrays.asList("100", "110", "120", "121", "240", "243", "400", "410", "600", "610", "700", "710", "720", "721", "800", "810", "900", "910", "980", "981");
 
-  public static final List<String> NAMES_X = new ArrayList<>(Arrays.asList("400", "410", "411", "700", "710", "711", "720", "721", "722", "900", "910", "911", "980", "981", "982"));
+  public static final List<String> NAMES_X = Arrays.asList("400", "410", "411", "700", "710", "711", "720", "721", "722", "900", "910", "911", "980", "981", "982");
 
-  public static final List<String> NAMES_V = new ArrayList<>(Arrays.asList("400", "410", "411", "800", "810", "811", "980", "981", "982"));
+  public static final List<String> NAMES_V = Arrays.asList("400", "410", "411", "800", "810", "811", "980", "981", "982");
 
-  public static final List<String> NAMES_245 = new ArrayList<>(Arrays.asList("600", "610", "611", "700", "710", "711", "720", "721", "722", "800", "810", "811", "900", "910", "911", "980", "981", "982"));
+  public static final List<String> NAMES_245 = Arrays.asList("600", "610", "611", "700", "710", "711", "720", "721", "722", "800", "810", "811", "900", "910", "911", "980", "981", "982");
 
-  public static final List<String> SUBJECTS_4 =  new ArrayList<>(Arrays.asList("600", "610", "611"));
+  public static final List<String> SUBJECTS_4 =  Arrays.asList("600", "610", "611");
 
-  public static final List<String> SUBJECTS_E =  new ArrayList<>(Arrays.asList("600", "610"));
+  public static final List<String> SUBJECTS_E =  Arrays.asList("600", "610");
 
-  public static final List<String> TITLES_X = new ArrayList<>(Arrays.asList("440", "730", "740", "930", "983"));
+  public static final List<String> TITLES_X = Arrays.asList("440", "730", "740", "930", "983");
 
-  public static final List<String> TITLES_V = new ArrayList<>(Arrays.asList("440", "830", "983"));
+  public static final List<String> TITLES_V = Arrays.asList("440", "830", "983");
 
-  public static final List<String> TITLES = new ArrayList<>(Arrays.asList("130", "241", "245", "730", "740", "830", "930", "941", "945", "983"));
+  public static final List<String> TITLES = Arrays.asList("130", "241", "245", "730", "740", "830", "930", "941", "945", "983");
 
   public static final String TERMINAL_PUNCTUATION = ".?!)-";
