+++ src/main/java/org/folio/marccat/dao/persistence/CasCache.java
@@ -5,7 +5,7 @@ import net.sf.hibernate.Session;
 import org.folio.marccat.business.common.Persistence;
 import org.folio.marccat.business.common.PersistenceState;
 import org.folio.marccat.dao.AbstractDAO;
-import org.folio.marccat.dao.CasCacheDAO;
+import org.folio.marccat.dao.MarcatCacheDAO;
 import org.folio.marccat.exception.DataAccessException;
 
 import java.io.Serializable;
@@ -387,7 +387,7 @@ public class CasCache implements Persistence {
 
   @Override
   public AbstractDAO getDAO() {
-    return new CasCacheDAO();
+    return new MarcatCacheDAO();
   }
 
   public String getNtrLevel() {
