+++ src/main/java/org/folio/marccat/integration/MarccatHelper.java
@@ -121,7 +121,7 @@ public abstract class MarccatHelper {
       .create()
       .username(config.get("user"))
       .password(config.get("password"))
-      .url(config.get("url"))
+      .url("jdbc:postgresql://192.168.0.158:5433/folio_marccat_sv2")
       .build();
   }
