--- src/main/resources/org/folio/marccat/dao/persistence/AUT_NTE.hbm.xml
@@ -1,19 +1,0 @@
-<?xml version="1.0"?>
-<!DOCTYPE hibernate-mapping PUBLIC 
-	"-//Hibernate/Hibernate Mapping DTD 2.0//EN"
-	"http://hibernate.sourceforge.net/hibernate-mapping-2.0.dtd">
-<hibernate-mapping package="librisuite.business.cataloguing.authority">
-
-	<class name="AuthorityNote" table="AUT_NTE">
-		<composite-id>
-			<key-property name="noteNumber" column="AUT_NTE_NBR"/>
-			<key-property name="itemNumber" column="AUT_NBR"/>
-		</composite-id>
-		<property name="noteStringText">
-			<column name="AUT_NTE_STRNG_TXT"/>
-		</property>
-		<property name="noteType">
-			<column name="AUT_NTE_TYP_CDE"/>
-		</property>
-	</class>
-</hibernate-mapping>
