--- src/main/resources/org/folio/marccat/dao/persistence/BND_CPY.hbm.xml
@@ -1,106 +1,0 @@
-<?xml version="1.0"?>
-<!DOCTYPE hibernate-mapping PUBLIC 
-	"-//Hibernate/Hibernate Mapping DTD 2.0//EN"
-	"http://hibernate.sourceforge.net/hibernate-mapping-2.0.dtd">
-<hibernate-mapping package="org.folio.marccat.dao.persistence">
-	<class name="BND_CPY" table="BND_CPY">
-		<id name="copyIdNumber">
-			<column name="CPY_ID_NBR" not-null="true"/>
-			<generator class="assigned"/>
-		</id>
-		<property name="bibItemNumber">
-			<column name="BIB_ITM_NBR" sql-type="NUMBER (11)" not-null="true"/>
-		</property>
-		<property name="shelfListKeyNumber">
-			<column name="SHLF_LIST_KEY_NBR" sql-type="NUMBER (11)"/>
-		</property>
-		<property name="organisationNumber">
-			<column name="ORG_NBR" sql-type="NUMBER (11)" not-null="true"/>
-		</property>
-		<property name="branchOrganisationNumber">
-			<column name="BRNCH_ORG_NBR" sql-type="NUMBER (11)" not-null="true"/>
-		</property>
-		<property name="originalOrganisationNumber">
-			<column name="ORGNL_ORG_NBR" sql-type="NUMBER (11)"/>
-		</property>
-		<property name="barCodeNumber">
-			<column name="BRCDE_NBR" sql-type="CHAR (14)"/>
-		</property>
-		<property name="dynixSerialIdNumber">
-			<column name="DYNIX_SRL_ID_NBR" sql-type="NUMBER (11)"/>
-		</property>
-		<property name="transactionDate">
-			<column name="TRSTN_DTE" sql-type="DATE" not-null="true"/>
-		</property>
-		<property name="creationDate">
-			<column name="CRTN_DTE" sql-type="DATE"/>
-		</property>
-		<property name="illCode">
-			<column name="ILL_CDE" sql-type="CHAR (1)" not-null="true"/>
-		</property>
-		<property name="holdingSubscriptionStatusCode">
-			<column name="HLDG_SBCPT_STUS_CDE" sql-type="CHAR (1)" not-null="true"/>
-		</property>
-		
-		<property name="holdingRetentionCode">
-			<column name="HLDG_RTNTN_CDE" sql-type="CHAR (1)" not-null="true"/>
-		</property>
-		<property name="loanPrd">
-			<column name="LOAN_PRD_CDE" sql-type="CHAR (1)" not-null="true"/>
-		</property>
-		<property name="holdingSeriesTrmtCode">
-			<column name="HLDG_SRS_TRMT_CDE" sql-type="CHAR (1)" not-null="true"/>
-		</property>
-		<property name="holdingStatusTypeCode">
-			<column name="HLDG_STUS_TYP_CDE" sql-type="CHAR (1)" not-null="true"/>
-		</property>
-		<property name="locationNameCode">
-			<column name="LCTN_NME_CDE" sql-type="NUMBER (6)" not-null="true"/>
-		</property>
-		<property name="holdingLevelOfDetailCode">
-			<column name="HLDG_LVL_OF_DTL_CDE" sql-type="CHAR (1)" not-null="true"/>
-		</property>
-		<property name="holdingAcsnListCode">
-			<column name="HLDG_ACSN_LIST_CDE" sql-type="CHAR (30)"/>
-		</property>
-		<property name="copyNumberDescription">
-			<column name="CPY_NBR_DSC" sql-type="VARCHAR2 (384)"/>
-		</property>
-		<property name="copyRemarkNote">
-			<column name="CPY_RMRK_NTE" sql-type="VARCHAR2 (768)"/>
-		</property>
-		<property name="copyStatementText">
-			<column name="CPY_STMT_TXT" sql-type="VARCHAR2 (2304)"/>
-		</property>
-		<property name="copyRemarkNoteSortForm">
-			<column name="CPY_RMRK_NTE_SRT_FORM" sql-type="VARCHAR2 (768)"/>
-		</property>
-		<property name="tempLocationOrganizationNumber">
-			<column name="TMP_LCTN_ORG_NBR" sql-type="NUMBER (11)"/>
-		</property>
-		<property name="tempLocationNameCode">
-			<column name="TMP_LCTN_NME_CDE" sql-type="NUMBER (6)"/>
-		</property>
-		<property name="materialDescription">
-			<column name="MTRL_DESC" sql-type="VARCHAR2 (256)"/>
-		</property>
-		<property name="cost">
-			<column name="CST" sql-type="NUMBER (15,6)"/>
-		</property>
-		<property name="currencyTypeCode">
-			<column name="CURCY_TYP_CDE" sql-type="NUMBER (6)"/>
-		</property>
-		<property name="currencyExchangeRte">
-			<column name="CURCY_XCHNG_RTE" sql-type="FLOAT"/>
-		</property>
-		<property name="transferCstdyNumber">
-			<column name="TRSFR_CSTDY_NBR" sql-type="NUMBER (11)"/>
-		</property>
-		<property name="physicalCopyType">
-			<column name="PHSCL_CPY_TPE" sql-type="NUMBER (1)"/>
-		</property>
-		<property name="methodAdquisition">
-			<column name="MTHD_ACQ" sql-type="VARCHAR2 (1)"/>
-		</property>
-	</class>
-</hibernate-mapping>
