--- src/main/resources/org/folio/marccat/dao/persistence/BRWR.hbm.xml
@@ -1,68 +1,0 @@
-<?xml version="1.0"?>
-<!DOCTYPE hibernate-mapping PUBLIC 
-	"-//Hibernate/Hibernate Mapping DTD 2.0//EN"
-	"http://hibernate.sourceforge.net/hibernate-mapping-2.0.dtd">
-<hibernate-mapping package="org.folio.marccat.dao.persistence">
- 	<class name="BRWR" table="BRWR">
-		<id name="borrowerNumber">
-			<column name="BRWR_NBR" not-null="true"/>
-			<generator class="native">
-				<param name="sequence">BRWR_SEQUENCE</param>
-			</generator>	
-		</id>
-		<property name="personNumber">
-			<column name="PRSN_NBR" sql-type="NUMBER (11)" not-null="true"/>
-		</property>
-		<property name="borrowerBarcodeNumber">
-			<column name="BRWR_BRCDE_NBR" sql-type="CHAR (14)" not-null="true"/>
-		</property>
-		<property name="borrowerLocalNumber">
-			<column name="BRWR_LCL_NBR" sql-type="NUMBER (11)"/>
-		</property>
-		<property name="personLocal1stTypeCode">
-			<column name="PRSN_LCL_1ST_TYP_CDE"  sql-type="NUMBER(6)"/>
-		</property>
-		<property name="personLocal2ndTypeCode">
-			<column name="PRSN_LCL_2ND_TYP_CDE" sql-type="NUMBER(6)"/>
-		</property>
-		<property name="personLocal3rdTypeCode">
-			<column name="PRSN_LCL_3RD_TYP_CDE" sql-type="NUMBER(6)"/>
-		</property>
-		<property name="personLocal4thTypeCode">
-			<column name="PRSN_LCL_4TH_TYP_CDE" sql-type="NUMBER(6)"/>
-		</property>
-		<property name="borrowerTypeCode">
-			<column name="BRWR_TYP_CDE" sql-type="NUMBER(6)"/>
-		</property>
-		<property name="borrowerUpdateTimeStamp">
-			<column name="BRWR_UPDT_TMEST" sql-type="DATE"  not-null="true"/>
-		</property>
-		<property name="borrowerBlacklistedDate">
-			<column name="BRWR_BLCKL_DTE" sql-type="DATE"/>
-		</property>
-		<property name="borrowerDeliveryIndicator">
-			<column name="BRWR_DLVRY_IND" sql-type="CHAR (1)"  not-null="true"/>
-		</property>
-		<property name="borrowerBorrowIndicator">
-			<column name="BRWR_BRW_IND" sql-type="CHAR (1)"  not-null="true"/>
-		</property>
-		<property name="borrowerTrapIndicator">
-			<column name="BRWR_TRP_IND" sql-type="CHAR (1)" not-null="true"/>
-		</property>
-		<property name="borrowerSummaryText">
-			<column name="BRWR_SMRY_TXT" sql-type="VARCHAR2 (60)"/>
-		</property>
-		<property name="organisationNumber">
-			<column name="ORG_NBR" sql-type="NUMBER (11)" not-null="true"/>
-		</property>
-		<property name="circulationDate">
-			<column name="CIR_DTE" sql-type="DATE"/>
-		</property>
-		<property name="circulationEndDate">
-			<column name="CIR_END_DATE" sql-type="DATE"/>
-		</property>
-		<property name="branchNumber">
-			<column name="ORG_BRNCH_NBR" sql-type="NUMBER (11)" not-null="true"/>
-		</property>
-	</class>
-</hibernate-mapping>
