--- src/main/resources/org/folio/marccat/dao/persistence/BRWR_FINE.hbm.xml
@@ -1,56 +1,0 @@
-<?xml version="1.0"?>
-<!DOCTYPE hibernate-mapping PUBLIC 
-	"-//Hibernate/Hibernate Mapping DTD 2.0//EN"
-	"http://hibernate.sourceforge.net/hibernate-mapping-2.0.dtd">
-<hibernate-mapping package="org.folio.marccat.dao.persistence">
- 	<class name="BRWR_FINE" table="BRWR_FINE">
-		<id name="borrowerFineNumber">
-			<column name="BRWR_FINE_NBR" not-null="true"/>
-			<generator class="native">
-				<param name="sequence">BRWR_FINE_SEQUENCE</param>
-			</generator>	
-		</id>
-		<property name="borrowerNumber">
-			<column name="BRWR_NBR" sql-type="NUMBER (11)" not-null="true"/>
-		</property>
-		<property name="copyIdNumber">
-			<column name="CPY_ID_NBR" sql-type="NUMBER (11)" not-null="true"/>
-		</property>
-		<property name="loanDueDate">
-			<column name="LOAN_DUE_DTE" sql-type="DATE" not-null="true"/>
-		</property>
-		<property name="loanChargeOutDate">
-			<column name="LOAN_CHRG_OUT_DTE" sql-type="DATE" not-null="true"/>
-		</property>
-		<property name="loanPreviousDueDate">
-			<column name="LOAN_PRVS_DUE_DTE" sql-type="DATE"/>
-		</property>
-		<property name="loanRecallNotificationDate">
-			<column name="LOAN_RCALL_NTFCT_DTE" sql-type="DATE"/>
-		</property>
-		<property name="loanFirstOverdueDate">
-			<column name="LOAN_1ST_OVRD_DTE" sql-type="DATE"/>
-		</property>
-		<property name="loanSecondOverdueDate">
-			<column name="LOAN_2ND_OVRD_DTE" sql-type="DATE"/>
-		</property>
-		<property name="loanThirdOverdueDate">
-			<column name="LOAN_3RD_OVRD_DTE" sql-type="DATE"/>
-		</property>
-		<property name="borrowerFineCreationDate">
-			<column name="BRWR_FINE_CRTE_DTE" sql-type="DATE" not-null="true"/>
-		</property>
-		<property name="borrowerFineFinalPaymentDate">
-			<column name="BRWR_FINE_FNL_PAYMT_DTE" sql-type="DATE"/>
-		</property>
-		<property name="loanRenewalCounterCnt">
-			<column name="LOAN_RNWL_CNTER_CNT" sql-type="NUMBER (11)"  not-null="true"/>
-		</property>
-		<property name="borrowerFineBlanceDueAmount">
-			<column name="BRWR_FINE_BLNC_DUE_AMT" sql-type="NUMBER (15,6)" not-null="true"/>
-		</property>
-		<property name="borrowerFineTotalFineAmount">
-			<column name="BRWR_FINE_TOTL_FINE_AMT" sql-type="NUMBER (15,6)"  not-null="true"/>
-		</property>
-	</class>
-</hibernate-mapping>
