--- src/main/resources/org/folio/marccat/dao/persistence/BRWR_FINE_PAYMT.hbm.xml
@@ -1,27 +1,0 @@
-<?xml version="1.0"?>
-<!DOCTYPE hibernate-mapping PUBLIC 
-	"-//Hibernate/Hibernate Mapping DTD 2.0//EN"
-	"http://hibernate.sourceforge.net/hibernate-mapping-2.0.dtd">
-<hibernate-mapping package="org.folio.marccat.dao.persistence">
- 	<class name="BRWR_FINE_PAYMT" table="BRWR_FINE_PAYMT">
-		<id name="borrowerFineNumber">
-			<column name="BRWR_FINE_NBR" not-null="true"/>
-			<generator class="assigned"/>
-		</id>
-		<property name="borrowerFinePaymentTimeStamp">
-			<column name="BRWR_FINE_PAYMT_TMEST" sql-type="DATE" not-null="true"/>
-		</property>
-		<property name="borrowerNumber">
-			<column name="BRWR_NBR" sql-type="NUMBER (11)" not-null="true"/>
-		</property>
-		<property name="borrowerFinePaymentAmount">
-			<column name="BRWR_FINE_PAYMT_AMT" sql-type="NUMBER (15,2)" not-null="true"/>
-		</property>
-		<property name="finePaymentAdjmtIndicator">
-			<column name="FINE_PAYMT_ADJMT_IND" sql-type="CHAR (1)" not-null="true"/>
-		</property>
-		<property name="finePaymentAdjmtNote">
-			<column name="FINE_PAYMT_ADJMT_NTE" sql-type="VARCHAR2 (48)"/>
-		</property>
-	</class>
-</hibernate-mapping>
