--- src/main/resources/org/folio/marccat/dao/persistence/BRWR_NTE.hbm.xml
@@ -1,15 +1,0 @@
-<?xml version="1.0"?>
-<!DOCTYPE hibernate-mapping PUBLIC 
-	"-//Hibernate/Hibernate Mapping DTD 2.0//EN"
-	"http://hibernate.sourceforge.net/hibernate-mapping-2.0.dtd">
-<hibernate-mapping package="org.folio.marccat.dao.persistence">
-	<class name="BRWR_NTE" table="BRWR_NTE">
-		<composite-id name="key" class="BorrowerNoteKey">
-			<key-property name="borrowerNumber" column="BRWR_NBR" />
-			<key-property name="borrowerNoteNumber" column="BRWR_NTE_NBR"/>
-		</composite-id>
-		<property name="noteText">
-			<column name="BRWR_NTE_CNTNT_DSC" sql-type="VARCHAR2 (384)" not-null="true"/>
-		</property>
-	</class>
-</hibernate-mapping>
