--- src/main/resources/org/folio/marccat/dao/persistence/BRWR_PSTL_ADR.hbm.xml
@@ -1,89 +1,0 @@
-<?xml version="1.0"?>
-<!DOCTYPE hibernate-mapping PUBLIC 
-	"-//Hibernate/Hibernate Mapping DTD 2.0//EN"
-	"http://hibernate.sourceforge.net/hibernate-mapping-2.0.dtd">
-<hibernate-mapping package="org.folio.marccat.dao.persistence">
- 	<class name="BRWR_PSTL_ADR" table="BRWR_PSTL_ADR">
- 		
- 		<composite-id name="key" class="BorrowerPostalAdrKey">
-			<key-property name="borrowerNumber" column="BRWR_NBR" />
-			<key-property name="personalPostalAddressTypeCode" column="PRSN_PSTL_ADR_TYP_CDE"/>
-		</composite-id>
- 		
-		<property name="postalAddressCityShortForm">
-			<column name="PSTL_ADR_CTY_SRT_FORM" sql-type="VARCHAR2 (96)"/>
-		</property>
-		<property name="postalAddressStreetNumber">
-			<column name="PSTL_ADR_ST_NBR" sql-type="CHAR (6)"/>
-		</property>
-		<property name="postalAddressFloorNumber">
-			<column name="PSTL_ADR_FLR_NBR" sql-type="CHAR (6)"/>
-		</property>
-		<property name="postalAddressRoomNumber">
-			<column name="PSTL_ADR_RM_NBR" sql-type="CHAR (6)"/>
-		</property>
-		<property name="postalAddressUnitDsgtrNumber">
-			<column name="PSTL_ADR_UNT_DSGTR_NBR" sql-type="CHAR (6)"/>
-		</property>
-		<property name="postalAddressDeliveryDsgtrNumber">
-			<column name="PSTL_ADR_DLVRY_DSGTR_NBR" sql-type="VARCHAR2 (20)"/>
-		</property>
-		<property name="postalAddressMdeCode">
-			<column name="PSTL_ADR_MDE_CDE" sql-type="NUMBER (6)"/>
-		</property>
-		<property name="postalAddressStreetTypeCode">
-			<column name="PSTL_ADR_ST_TYP_CDE" sql-type="NUMBER (6)"/>
-		</property>
-		<property name="postalAddressStreetDirectionCode">
-			<column name="PSTL_ADR_ST_DRCTN_CDE" sql-type="NUMBER(6)"/>
-		</property>
-		<property name="postalAddressUnitDsgtrCode">
-			<column name="PSTL_ADR_UNT_DSGTR_CDE" sql-type="NUMBER(6)"/>
-		</property>
-		<property name="postalAddressDeliveryDsgtrCode">
-			<column name="PSTL_ADR_DLVRY_DSGTR_CDE" sql-type="NUMBER(6)"/>
-		</property>
-		<property name="postalAddressIntlnTypeCode">
-			<column name="PSTL_ADR_INTLN_TYP_CDE" sql-type="NUMBER(6)"/>
-		</property>
-		<property name="postalAddressDepartmentNameCode">
-			<column name="PSTL_ADR_DEPT_NME_CDE" sql-type="CHAR (2)"/>
-		</property>
-		<property name="postalAddressRegionCode">
-			<column name="PSTL_ADR_RGN_CDE" sql-type="NUMBER(6)"/>
-		</property>
-		<property name="postalAddressGeographyCode">
-			<column name="PSTL_ADR_GEOG_CDE" sql-type="NUMBER(6)"/>
-		</property>
-		<property name="postalAddressProvincialTerritoryStateCode">
-			<column name="ADR_PROV_TRTRY_STATE_CDE" sql-type="CHAR (2)"/>
-		</property>
-		<property name="postalAddressPostalCode">
-			<column name="PSTL_ADR_PSTL_CDE" sql-type="VARCHAR2 (20)"/>
-		</property>
-		<property name="postalAddressCountryCode">
-			<column name="PSTL_ADR_CNTRY_CDE" sql-type="NUMBER(6)"/>
-		</property>
-		<property name="postalAddressIntlnQualigyDescriptor">
-			<column name="PSTL_ADR_INTLN_QLFY_DSC" sql-type="VARCHAR2 (60)"/>
-		</property>
-		<property name="postalAddressReferenceDescriptor">
-			<column name="PSTL_ADR_REF_DSC" sql-type="VARCHAR2 (96)"/>
-		</property>
-		<property name="postalAddressPersonalAttentionName">
-			<column name="PSTL_ADR_PRSN_ATTN_NME" sql-type="VARCHAR2 (768)"/>
-		</property>
-		<property name="postalAddressOrganisationName">
-			<column name="PSTL_ADR_ORG_NME" sql-type="VARCHAR2 (96)"/>
-		</property>
-		<property name="postalAddressStreetName">
-			<column name="PSTL_ADR_ST_NME" sql-type="VARCHAR2 (96)"/>
-		</property>
-		<property name="postalAddressBldgName">
-			<column name="PSTL_ADR_BLDG_NME" sql-type="VARCHAR2 (96)"/>
-		</property>
-		<property name="postalAddressCityName">
-			<column name="PSTL_ADR_CTY_NME" sql-type="VARCHAR2 (96)"/>
-		</property>
-	</class>
-</hibernate-mapping>
