--- src/main/resources/org/folio/marccat/dao/persistence/CAS_ESITO_IMPORT.hbm.xml
@@ -1,18 +1,0 @@
-<?xml version="1.0"?>
-<!DOCTYPE hibernate-mapping PUBLIC 
-	"-//Hibernate/Hibernate Mapping DTD 2.0//EN"
-	"http://hibernate.sourceforge.net/hibernate-mapping-2.0.dtd">
-<hibernate-mapping package="org.folio.marccat.dao.persistence">
-	<class name="CasEsitoImport" table="CAS_ESITO_IMPORT">	
-		<id name="transactionId">
-			<column name="TRANSACTION_ID" not-null="true"/>
-			<generator class="assigned"/>
-		</id>	
-		<property name="idFonte">
-			<column name="FONTE" sql-type="VARCHAR2 (2)"/>
-		</property>	
-		<property name="idEsito">
-			<column name="ID_ESITO" sql-type="VARCHAR2 (1)"/>
-		</property>		
-	</class>
-</hibernate-mapping>
