--- src/main/resources/org/folio/marccat/dao/persistence/CAS_FILTRI_MIGRAZIONE.hbm.xml
@@ -1,17 +1,0 @@
-<?xml version="1.0"?>
-<!DOCTYPE hibernate-mapping PUBLIC 
-	"-//Hibernate/Hibernate Mapping DTD 2.0//EN"
-	"http://hibernate.sourceforge.net/hibernate-mapping-2.0.dtd">
-<hibernate-mapping package="org.folio.marccat.dao.persistence">
-	<class name="CasFiltriMigrazione" table="CAS_FILTRI_MIGRAZIONE">	
-		<composite-id>
-	        <key-property name="idFonte" column="FONTE"/>
-	        <key-property name="filtro" column="FILTRO"/>
-	    </composite-id>  
-	
-	
-		<property name="ragione">
-			<column name="RAGIONE" sql-type="VARCHAR2 (100)"/>
-		</property>
-	</class>
-</hibernate-mapping>
