--- src/main/resources/org/folio/marccat/dao/persistence/CAS_FLTR_IMP_SPOT.hbm.xml
@@ -1,23 +1,0 @@
-<?xml version="1.0"?>
-<!DOCTYPE hibernate-mapping PUBLIC 
-	"-//Hibernate/Hibernate Mapping DTD 2.0//EN"
-	"http://hibernate.sourceforge.net/hibernate-mapping-2.0.dtd">
-<hibernate-mapping package="org.folio.marccat.dao.persistence">
-	<class name="CasFiltriImportSpot"  schema="custom"  table="CAS_FLTR_IMP_SPOT">	
-		<composite-id>
-	        <key-property name="idFonte" column="SRC_ID"/>
-	        <key-property name="idFiltro" column="FLTR_ID"/>
-	    </composite-id>	
-		<property name="valFiltro">
-			<column name="FILTER" sql-type="VARCHAR2 (128)"/>
-		</property>
-		<!-- <joined-subclass
-				name="CasFiltri"
-				table="T_CAS_FLTR">
-			<key column="FLTR_ID"/>
-			<property name="idFiltro" column="FLTR_ID"/>
-			<property name="nomeFiltro" column="DESCR"/>
-			<property name="tipoFiltro" column="FLTR_TYP"/>
-		</joined-subclass> -->
-	</class>
-</hibernate-mapping>
