--- src/main/resources/org/folio/marccat/dao/persistence/CasDataExp.hbm.xml
@@ -1,37 +1,0 @@
-<?xml version="1.0"?>
-<!DOCTYPE hibernate-mapping PUBLIC 
-	"-//Hibernate/Hibernate Mapping DTD 2.0//EN"
-	"http://hibernate.sourceforge.net/hibernate-mapping-2.0.dtd">
-    
-<hibernate-mapping>
-
-<class 
-    name="org.folio.marccat.dao.persistence.CasDataExp"
-     schema="custom" 
-    table="CAS_DATA_EXP"
-    lazy="false"
->
-	<id
-        name="bibItmNbr"
-        type="java.lang.Long"
-        column="BIB_ITM_NBR"
-    >
-     <generator class="assigned" />
-    </id>
-
-	<property name="srcId">
-		<column name="SRC_ID" sql-type="NUMBER (2)"/>
-	</property>
-	
-	<property name="impDte">
-		<column name="IMP_DTE" sql-type="DATE"/>
-	</property>
-	
-	<property name="invcNbr">
-		<column name="INVC_NBR" sql-type="VARCHAR2 (10)"/>
-	</property>
-</class>
-
-</hibernate-mapping>
-
-
