--- src/main/resources/org/folio/marccat/dao/persistence/CasDataNTI.hbm.xml
@@ -1,40 +1,0 @@
-<?xml version="1.0"?>
-<!DOCTYPE hibernate-mapping PUBLIC 
-	"-//Hibernate/Hibernate Mapping DTD 2.0//EN"
-	"http://hibernate.sourceforge.net/hibernate-mapping-2.0.dtd">
-    
-<hibernate-mapping>
-
-<class 
-    name="org.folio.marccat.dao.persistence.CasDataNTI"
-     schema="custom" 
-    table="CAS_DATA_NTI"
-    lazy="false"
->
-	<id
-        name="bibItmNbr"
-        type="java.lang.Long"
-        column="BIB_ITM_NBR"
-    >
-     <generator class="assigned" />
-    </id>
-    
-    <property name="transactionId">
-		<column name="TRANSACTION_ID" sql-type="VARCHAR2 (40)"/>
-	</property>
-	
-	<property name="operator">
-		<column name="OPERATOR" sql-type="VARCHAR2 (20)"/>
-	</property>
-	
-	<property name="ntiDate">
-		<column name="NTI_DATE" sql-type="DATE"/>
-	</property>
-	
-	
-</class>
-
-</hibernate-mapping>
-
-
-
