--- src/main/resources/org/folio/marccat/dao/persistence/CasDgaPolicy.hbm.xml
@@ -1,30 +1,0 @@
-<?xml version="1.0"?>
-<!DOCTYPE hibernate-mapping PUBLIC
-    "-//Hibernate/Hibernate Mapping DTD 2.0//EN"
-    "http://hibernate.sourceforge.net/hibernate-mapping-2.0.dtd" >
-    
-<hibernate-mapping package="com.casalini.hibernate.model">
-<class name="CasDgaPolicy"  schema="custom"  table="CAS_DGA_POLICY">     
-     <composite-id>
-        <key-property name="bibItemNumber" column="BIB_ITM_NBR"/>
-        <key-property name="idPolicy" column="ID_PLC"/>
-        <key-property name="typePolicy" column="TYPE_PLC"/>
-     </composite-id>
-     
-    <property name="denPolicy"> 
-		<column name="DEN_PLC"/>
-	</property>
-    <property name="pricePolicy"> 
-		<column name="PRICE_PLC"/>
-	</property>
-    <property name="numStampaPolicy"> 
-		<column name="NUM_OPE_STAMPA_PLC"/>
-	</property>
-    <property name="curcyPolicy"> 
-		<column name="CURCY_PLC"/>
-	</property>
-	<property name="priceTotal"> 
-		<column name="PRICE_TOTAL"/>
-	</property>	
-</class>
-</hibernate-mapping>
