--- src/main/resources/org/folio/marccat/dao/persistence/CasDigFascicoli.hbm.xml
@@ -1,64 +1,0 @@
-<?xml version="1.0"?>
-<!DOCTYPE hibernate-mapping PUBLIC
-    "-//Hibernate/Hibernate Mapping DTD 3.0//EN"
-    "http://hibernate.sourceforge.net/hibernate-mapping-3.0.dtd" >
-    
-<hibernate-mapping>
-<!-- 
-    Created by the Middlegen Hibernate plugin 2.2
-
-    http://boss.bekk.no/boss/middlegen/
-    http://www.hibernate.org/
--->
-
-<class 
-    name="org.folio.marccat.dao.persistence.CasDigFascicoli"
-    table="CAS_DIG_FASCICOLI"
-    lazy="false"
->
-
-    <id
-        name="idFascicolo"
-        type="java.lang.Long"
-        column="ID_FASCICOLO"
-    >
-    
-        <generator class="assigned" />
-    </id>
-
-    <property
-        name="bibItmNbrMadre"
-        type="java.lang.Long"
-        column="BIB_ITM_NBR_MADRE"
-        length="11"
-    />
-    <property
-        name="bibItmNbrFiglia"
-        type="java.lang.Long"
-        column="BIB_ITM_NBR_FIGLIA"
-        length="11"
-    />
-    <property
-        name="anno"
-        type="java.lang.String"
-        column="ANNO"
-        length="20"
-    />
-    <property
-        name="volume"
-        type="java.lang.String"
-        column="VOLUME"
-        length="20"
-    />
-    <property
-        name="fascicolo"
-        type="java.lang.String"
-        column="FASCICOLO"
-        length="20"
-    />
-
-    <!-- Associations -->
-  
-
-</class>
-</hibernate-mapping>
