--- src/main/resources/org/folio/marccat/dao/persistence/CasEresAdmin.hbm.xml
@@ -1,121 +1,0 @@
-<?xml version="1.0"?>
-<!DOCTYPE hibernate-mapping PUBLIC
-    "-//Hibernate/Hibernate Mapping DTD 2.0//EN"
-    "http://hibernate.sourceforge.net/hibernate-mapping-2.0.dtd" >
-    
-<hibernate-mapping package="com.casalini.hibernate.model">
-<class name="CasEresourceAdmin"  schema="custom"  table="CAS_ERES_ADMIN">
-    <composite-id>
-        <key-property name="bibItemNumber" 
-            column="BIB_ITM_NBR"/>
-     </composite-id>    
-    <property name="fileName"> 
-		<column name="FILE_NAME"/>
-	</property>
-   <property name="fileSize"> 
-		<column name="FILE_SIZE"/>
-	</property>
-	<property name="flagWebStatus"> 
-		<column name="FLAG_WEB_STATUS"/>
-	</property>
-	<property name="flagAllegato"> 
-		<column name="FLAG_ALLEGATO"/>
-	</property>
-	<property name="flagSupplemento"> 
-		<column name="FLAG_SUPPLEMENTO"/>
-	</property>
-	<property name="flagStato"> 
-		<column name="FLAG_STATO"/>
-	</property>
-	<property name="flagCopiaIncolla"> 
-		<column name="FLAG_COPIA_INCOLLA"/>
-	</property>
-	<property name="opeStampa"> 
-		<column name="FLAG_OPER_STAMPA"/>
-	</property>
-	<property name="numStampa"> 
-		<column name="NUM_OPE_STAMPA"/>
-	</property>
-	<property name="costoPagina"> 
-		<column name="COSTO_PAGINA"/>
-	</property>
-	<property name="costoPaginaCurcy"> 
-		<column name="COSTO_PAGINA_CURCY"/>
-	</property>
-	<property name="costoRivista"> 
-		<column name="COSTO_RIVISTA"/>
-	</property>
-	<property name="costoRivistaCurcy"> 
-		<column name="COSTO_RIVISTA_CURCY"/>
-	</property>
-	<property name="costoArticolo"> 
-		<column name="COSTO_ARTICOLO"/>
-	</property>
-	<property name="costoArticoloCurcy"> 
-		<column name="COSTO_ARTICOLO_CURCY"/>
-	</property>
-	<property name="costoSezione"> 
-		<column name="COSTO_SEZIONE"/>
-	</property>
-	<property name="costoSezioneCurcy"> 
-		<column name="COSTO_SEZIONE_CURCY"/>
-	</property>
-	<property name="costoCapitolo"> 
-		<column name="COSTO_CAPITOLO"/>
-	</property>
-	<property name="costoCapitoloCurcy"> 
-		<column name="COSTO_CAPITOLO_CURCY"/>
-	</property>
-	<property name="costoItem"> 
-		<column name="COSTO_ITEM"/>
-	</property>
-	<property name="costoItemCurcy"> 
-		<column name="COSTO_ITEM_CURCY"/>
-	</property>
-	<property name="pageTot"> 
-		<column name="NUM_PAGE_TOT"/>
-	</property>
-	<property name="saleTypeId"> 
-		<column name="SALE_TYPE_ID"/>
-	</property>
-	<property name="itemTypeSale"> 
-		<column name="ITEM_TYPE_SALE"/>
-	</property>
-	<property name="anno"> 
-		<column name="ANNO"/>
-	</property>
-	<property name="volume"> 
-		<column name="VOLUME"/>
-	</property>
-	<property name="idFascicolo"> 
-		<column name="ID_FASCICOLO"/>
-	</property>
-	<property name="capitolo"> 
-		<column name="CAPITOLO"/>
-	</property>
-	<property name="pagineDa"> 
-		<column name="PAGINE_DA"/>
-	</property>
-	<property name="pagineA"> 
-		<column name="PAGINE_A"/>
-	</property>
-	<property name="codEditore"> 
-		<column name="CODICE_EDITORE"/>
-	</property>
-	<property name="codEditoreBreve"> 
-		<column name="CODICE_EDITORE_BREVE"/>
-	</property>
-	<property name="denEditore"> 
-		<column name="DENOMINAZIONE_EDITORE"/>
-	</property>
-	<property name="progressivo"> 
-		<column name="PROGRESSIVO"/>
-	</property>
-	<property name="descrizione"> 
-		<column name="DESCRIZIONE"/>
-	</property>
-	<property name="pdf"> 
-		<column name="PDF"/>
-	</property>
-</class>
-</hibernate-mapping>
