--- src/main/resources/org/folio/marccat/dao/persistence/CasFiles.hbm.xml
@@ -1,47 +1,0 @@
-<?xml version="1.0"?>
-<!DOCTYPE hibernate-mapping PUBLIC
-    "-//Hibernate/Hibernate Mapping DTD 2.0//EN"
-    "http://hibernate.sourceforge.net/hibernate-mapping-2.0.dtd" >
-    
-<hibernate-mapping package="com.casalini.hibernate.model">
-<class name="CasFiles"  schema="custom"  table="CAS_FILES">
-
-	<id
-        name="idItem"
-        type="java.lang.Integer"
-        column="ID_ITEM"
-    >
-    	<generator class="native">
-			<param name="sequence">custom.ITEM_SEQ</param>
-		</generator>       
-    </id>
- 
-    <property name="bibItemNumberMadre"> 
-		<column name="BIB_ITM_NBR_MADRE"/>
-	</property>
-   <property name="bibItemNumberFiglia"> 
-		<column name="BIB_ITM_NBR_FIGLIA"/>
-	</property>
-	<property name="ntrLevel"> 
-		<column name="NTR_LVL"/>
-	</property>
-	<property name="orderProgr"> 
-		<column name="ORDER_NUM_PROGR"/>
-	</property>
-	<property name="cntlKeyNbr"> 
-		<column name="CNTL_KEY_NBR"/>
-	</property>
-	<property name="numTomo"> 
-		<column name="NUM_TOMO"/>
-	</property>
-	<property name="numEdizione"> 
-		<column name="NUM_EDIZ"/>
-	</property>
-	<property name="note"> 
-		<column name="NOTE"/>
-	</property>
-	<property name="progressiveType"> 
-		<column name="PROGR_TYPE"/>
-	</property>
-</class>
-</hibernate-mapping>
