--- src/main/resources/org/folio/marccat/dao/persistence/CasPublCdeIsbn.hbm.xml
@@ -1,24 +1,0 @@
-<?xml version="1.0"?>
-<!DOCTYPE hibernate-mapping PUBLIC
-    "-//Hibernate/Hibernate Mapping DTD 2.0//EN"
-    "http://hibernate.sourceforge.net/hibernate-mapping-2.0.dtd" >
-    
-<hibernate-mapping package="com.casalini.hibernate.model">
-<class name="CasPublCdeIsbn"  schema="custom"  table="CAS_PUBL_CDE_ISBN">
-    <composite-id>
-        <key-property name="isbnSortForm" column="ISBN_SRT_FORM"/>
-        <key-property name="codEditore" column="CODICE_EDITORE"/>
-     </composite-id>
-     
-	<property name="isbnStringText"> 
-		<column name="ISBN_STRNG_TX"/>
-	</property>	
-	<property name="user"> 
-		<column name="USR_NME"/>
-	</property>	
-	<property name="date"> 
-		<column name="DTE"/>
-	</property>	
-		
-</class>
-</hibernate-mapping>
