--- src/main/resources/org/folio/marccat/dao/persistence/CasQltyStus.hbm.xml
@@ -1,40 +1,0 @@
-<?xml version="1.0"?>
-<!DOCTYPE hibernate-mapping PUBLIC 
-	"-//Hibernate/Hibernate Mapping DTD 2.0//EN"
-	"http://hibernate.sourceforge.net/hibernate-mapping-2.0.dtd">
-    
-<hibernate-mapping>
-
-<class 
-    name="org.folio.marccat.dao.persistence.CasQltyStus"
-     schema="custom" 
-    table="CAS_QLTY_STUS"
-    lazy="false"
->
-	<id
-        name="transactionId"
-        type="java.lang.Integer"
-        column="TRANSACTION_ID"
-    >
-    	<generator class="native">
-			<param name="sequence">Casalini.TRANS_QUALITY_SEQ</param>
-		</generator>       
-    </id>
-	
-	<property name="operator">
-		<column name="USERNAME" sql-type="VARCHAR2 (100)"/>
-	</property>
-	
-	<property name="status">
-		<column name="STATUS" sql-type="CHAR(1)"/>
-	</property>
-
-	<property name="description">
-		<column name="DESCRIPTION" sql-type="VARCHAR2 (200)"/>
-	</property>	
-	
-</class>
-
-</hibernate-mapping>
-
-
