--- src/main/resources/org/folio/marccat/dao/persistence/OPEN_SUITE_RELEASE.hbm.xml
@@ -1,19 +1,0 @@
-<?xml version="1.0"?>
-<!DOCTYPE hibernate-mapping PUBLIC 
-	"-//Hibernate/Hibernate Mapping DTD 2.0//EN"
-	"http://hibernate.sourceforge.net/hibernate-mapping-2.0.dtd">
-<hibernate-mapping package="org.folio.marccat.dao.persistence">
-	<class name="OPEN_SUITE_RELEASE"  schema="OLISUITE"  table="WECAT_RELEASE">
-		 <composite-id>
-        <key-property 
-            name="servicePackNumber" 
-            column="SERVICE_PACK_NUMBER" 
-        />
-        <key-property 
-            name="releaseNumber" 
-            column="RELEASE_NUMBER" 
-        />
- 
-    </composite-id>  
-	</class>
-</hibernate-mapping>
